refactor: relocate canonicalizeEmail to test-fixtures as a test utility
canonicalizeEmail/CanonicalEmail have no production caller yet, so they no
longer belong in the production @packages/domain package as tested-but-unused
code. Move them to @packages/test-fixtures — the test-utility package whose
coverage is accommodated by usage — mirroring how this PR promotes
normalizeEmail into domain because production does consume it. They get
promoted back to domain by the later PR that wires them into the auth flow.

Also fixes the remaining low-priority review items on the moved code:
- Replace the negative .not.toBe assertion with two positive assertions, so
  the test proves the two off-Gmail addresses do not collapse by pinning each.
- Drop the raw email (PII) interpolated into the assert messages; the label
  alone identifies which invariant failed.

normalizeEmail stays in @packages/domain/user because production consumes it.
