+++ src/packages/domain/src/user/index.ts
@@ -1,6 +1,5 @@
 export type { UserId, AuthenticatedUserId } from "./user.types";
 export { UserIdSchema, authenticatedUserIdFrom } from "./user.schema";
-export type { CanonicalEmail } from "./email";
-export { canonicalizeEmail, normalizeEmail } from "./email";
+export { normalizeEmail } from "./email";
 export type { UserIdPrefix } from "./user-id-prefix";
 export { USER_ID_PREFIX_LENGTH, UserIdPrefixSchema, userIdPrefixFrom, parseUserIdPrefix } from "./user-id-prefix";
